Add RabbitMQ Plugin Support
These plugins are modeled using a defined resource type named
rabbitmq::plugin.  They notify the service and require the package.
However, this is not ideal since the defined type is establishing
relationships to the resources inside the rabbitmq class.

I plan to refactor rabbitmq into rabbitmq::service to address this
issue.  Then, I'll be able to use class level relationships.