Ensure nova_manage detects existing networks
Previously, there was a bug in the exists? method of
the nova_network provider where it could not detect
if a network already existed. This is b/c it
appended an extra /\d\d on the network id.

This commit corrects the logic so that existing
networks can be correctly detected.