Add proxy to apt class
This commit adds a proxy to the apt class in the
pre config manifest.

This is because I am using a squid proxy to speed
up testing in my environment.

I need to expose this as a config option so that
the hardcoded address of my app proxy is not
required for anyone to use this.