Add custom upstart scripts
In the current trunk PPA for oneiric (which is what
I am using for testing), the upstart files seem
to have a problem where they only will start the
swift servers if the file:
  /etc/swift/<server_name>-server

exists. Raised as issue: 931893
in Launchpad.

This commit adds puppet code to deploy custom
upstart scripts that have fixed this issue.