Install nova packages per service
This commit ensures that the individual Ubuntu nova
packages for any given service are only installed when that
service is configured.

This behavior was broken when RedHat support was added.
Since Redhat only has a single package, it was assumed
that all packages should always be installed on all nodes.

This was causing all services to be in a running state on
Ubuntu (b/c the packages were starting the related service)