Stop caching user/tenant/role lookups in user_role
Ensures that user tenant and role ids are always
looked up.

There was a previous issue where any users, tenant,
or roles created after any user_roles would not be
found b/c the results of querying for those things
were cached on the first user_role run.

This resolves the issue by not caching the results
of these lookups. I would like to explore figuring
out to do this in a more performant way.

It is possible that these kind of caching bugs still
exist.