Update parameters in nova class
This commit updates three parameters in the nova class.

- use_deprecated_auth

Previously, this was always set to be true. This commit updates
the config so that this is set to false if auth_type is keystone (and
set to true if it is not)

- glance_host, glance_port

These options have been removed since they had already been replaced with
the glance_api_servers option.