Perform explicit db migrations
The db sync script should be used to create
the initial db schema so that it will create
a migration revision table.

This migration revision table allows for schema upgrades.

This commit adds an exec resource for glance-manage db_sync
and ensures that it only occurs if the registry.conf file,
package, or db resource change.