Simplify keystone provider test
The previous test for the keystone parent provider that
tested that it failed correctly when no keystone config file
existed was too complicated. It was using some strange mocking
to try to mock a case that looked like the case of what would happen
when the file did not exist.

This commit changes the test not to use mocking which was way too complicated.
Instead, it uses tempfile to actually pass a file that did not exist.