Update nova_network type
This commits changes two things related to nova_network.

Adds vlan_start parameter. This parameter is neeed when the vlan being
created is not the first one that is created (b/c it cannot use the
vlan_start flag from nova.conf if the vlan already exists)

Removes the default openstack tenant from nova::manage::network so that the
vlans are automatically assigned to projects by default.