Update spec tests to work with ri's version of concat
This commit fixes spec test failures that were caused by
using RI's version of the concat module.

It introduces a custom fact concat_basedir that can determine
where the vardir is for a particular agent.

This commits adds that fact to unit tests that are validating
the resuling fragment files.