Commit spec tests for the haproxy module
Previously, the haproxy module didn't provide rspec tests for the class
or defined resource types.  This commit adds the tests, a spec_helper
file, and a rakefile so you can do `rake spec` and automatically run the
tests.

I've also committed a symlink in the
spec/fixtures/modules/haproxy directory so the rspec-puppet tests will
FIND the haproxy class/defined resource types, but I HAVE NOT committed
a symlink so that it will find the concat class (which is a dependency
for this module).  If you choose to run the tests, you must symlink the
concat module in the spec/fixtures/modules directory before the spec
tests will pass.