Add enabled parameter to glance services.
This commit adds the parameter enabled to both glance::api
as well as glance::registry.

Setting enabled to false ensures that the service is not running
and that the db schema is not create/migrated.

This flag is being added to all nova services to allow the deployment
of the openstack controller in active-passive for HA.