Update specs for haproxy base class with debian support
Many of the tests that apply to the RedHat family also apply to Debian.
The tests have been updated to reflect this, as well as to capture the
minor differences between the two platforms.

This commit incorperates the changes from glarizza#7 and glarizza#12

Not all tests pass because I found it pertinent to update the spec
where the actual module code was lacking.