Unconditionaly do pvcreate
During the disk partitioning pmanager runs pvcreate
which asks for confirmation on wiping out the volume
group if it is found there. There was another fix
which does dd to clean up all information about
all disk partionins, so pvcreate should not ask
anything. This patch adds '-y' option to pvcreate,
and it writes changes without asking questions,
so we can be double sure that installation won't
stuck under some other condition on this command.

Change-Id: I7368ce58fa54155664d59adedbfc0463f7d3a3ea