Finish configuring a service before adding it to HAProxy
Starting an HAProxy instance for ceilometer-api (and possibly other
similar services) will fail on Ubuntu if the package for that service is
installed but not yet configured: the service is still running in its
default configuration binding to 0.0.0.0 instead of just the management
IP, preventing HAProxy from binding to the management VIP on the same
port.

To prevent a dependency loop (Ceilometer, like most other OpenStack
services, depends on MySQL), Galera dependencies have to be narrowed
down from the whole of HAProxy setup to just the registration of MySQL
service with it. HAProxy related parts of controller_ha and galera
classes are refactored to support reloading HAProxy for each service
independently, and to group together HAProxy options specific to
individual services.

Added "--no-class_inherits_from_params_class-check" to puppet-lint
options to prevent it from issuing a warning that is not relevant for
Puppet versions 2.7 and newer.

Change-Id: I5fa03dbc3f7e3a7dfa57c226a5396f652c111997
Closes-bug: #1264388