Multiple Cluster Networks: add ip rule route
Adds support for ip route rules (ip route policy) to interfaces with
gateways supplied. This will create a policy based default route for
said interface. This is necessary to provide return routes when multiple
cluster networks are defined, otherwise packets would only return on the
default interface when the are outside of the network of the interface.

Adds disabled code to configure haproxy to keep vip points together
There are some deployment cases where the deployer may need to enable
this and this serves as the code example to perform it.

Resolves-bug: #1272766
Implements: bp multiple-cluster-networks
Change-Id: I06e09fd9b292cac53b925933006de064e725bdb0