Added ironic puppet module.
As far as we are moving toward closer openstack integration
as a first step we are planning to use ironic as our
provisioning tool. For the start we are planning to
avoid using ISCSI, glance, neutron so we need just ironic
and keystone installed on master node.
Puppet module which is able to deploy ironic from
source git repository. The work is in progress.
Will be improved in the future. For details look at
https://blueprints.launchpad.net/fuel/+spec/provision-ironic-puppet

Change-Id: I7ab5746be3ba773d623d925368bfecd0845575b6