Made mkfs arguments optional and added rootdelay
The problem is that we have been creating file
systems with -f flag while it is suitable only for
xfs, but not for any other file systems. Another
problem is that some hardware such as ancient Dell
equipment delays exposing raid volumes and grub
usually fails not waiting for root file system is
available. rootdelay=90 allows to avoid that kind of
unexpected behaviour.

Change-Id: I199077b4bf53fdcdc17a89fb4c8ad82bfd966a87