Fix RGW endpoint port
Change RGW endpoint port registered in keystone
to 8080 as this is the real port on which
haproxy is listening.

Previously, until https://review.openstack.org/#/c/85195/
was merged, RGW was working because virtual IP was located
on one of the controllers and apache proxying RGW fastcgi
application was answering on 6780 port. Currently, VIP moved
into namespace and 6780 is no longer available.

Change-Id: I2ad74b743561d21bb8e522c0648d26d93348657e
Closes-Bug: #1307508