Add log rotation for docker containers
Run a cronjob which executes a command like:

  dockerctl shell CONTAINER logrotate /etc/logrotate.conf

This will allow logrotate to reload needed services using configs
shipped with the packages.

We aslo need to remove "rsyslog" and "nginx" packages from
containers that use "nailgun::packages" class to avoid logrotate
duplication. So we create a new class "nailgun::common" and use it
in our $service-only.pp manifests instead of "nailgun::packages"

blueprint fuel-containerization-of-services

Change-Id: Ic625e086db7bcd3622bce73ec4cec0dd6f81121d