Let kernel kill dead TCP connections
1) Kill dead tcp connections in amore effective way:
configure the keepalive routines to wait for 5 secs before sending
the first keepalive probe, and then resend it every 5 seconds.
If no ACK response is received for 1 try, the connection is marked
as broken.
(The defaults are 7200, 75, 9 respectively and provide a *very* poor logic
for dead connections tracking and failover as well)
http://tldp.org/HOWTO/TCP-Keepalive-HOWTO/usingkeepalive.html
2) Set report_interval to 60 seconds and service_down_time to 180
seconds for Nova to let kernel kill dead connections.
3) Fix missing report_interval param usage in nova class.

Change-Id: Ic9d491f4904a5e665278027fc37254003c4b5172
Closes-Bug: #1289200