Add link to /root/.cephdeploy.conf -> /etc/ceph/ceph.conf
ceph-deploy requires that its local config file (previously $(pwd)/ceph.conf)
is exactly the same as /etc/ceph/ceph.conf on the system or --overwite-conf
has to be specified. Since we need to manipulate it and can't guarantee order
 from Puppet / Ruby we just link to two on the primary-controller to avoid
 the issue.

All additional users of config should config pull with --overwite-conf and
not have this issue.

The old link is left in place at this time to support older versions of
 ceph-deploy and can be removed in a later release.

Closes-bug: #1333814
Change-Id: I0b61c71effcf779801756968b8ed8b8cb93cb54b