RabbitMQ node resource level fencing
When Corosync notifies that particular node in its cluster
is dead, this script fences the failed node in RabbitMQ cluster
as well.
Note: system events may be monitored with dbus-monitor --system.
* It casts disconnect failed_node & forget_cluster_node for
the rest of the nodes in the RabbitMQ cluster.
* Does not fence alive nodes with mnesia running.
* Does not fence already forgotten nodes, that means that
  only the first node detected a 'dead event' will issue the
  fencing action, while the rest of the cluster nodes will
  ignore it.
* Requires corosync compiled with --enable-dbus option,
  ensures corosync-notifyd and dbus (messagebus) are running.
* Can be run as a daemon by manual start of rabbit-fence.py
  and is disabled by default.
TODO make package with init script for rabbit-fence daemon

Doc-Impact: ops guide
Change-Id: I691363386efe01421acc317ef6371ce45a0d4d11