create keystone endpoints after keystone service
Keystone endpoints are created using python-keystoneclient which uses Keystone
API, which means that keystone service should be already running. The
dependency to run keystone db_sync before *::keystone::auth is not strong
enough, so it is replaced with dependency on keystone service (which in turn is
subscribed to keystone db_sync).

Since nova_floating_range depends on nova::keystone::auth (in cluster_simple),
this will also ensure that keystone service is running before floating IP
ranges are created with nova-network.

Closes-Bug: #1348756
Change-Id: Iad05b005fcc99d50507c36e934bce1fd1937ab6c