WIP: Refactor rabbitmq OCF script
This refactoring adds improvements and fixes several
possible and already filed issues making rabbitmq
cluster reassembling in case of partial or complete
failure.

1) Use mnesia low-level commands instead of
status and cluster_status because these
commands will not block in case of
one of the nodes becoming inaccessible

2) Block access to rabbitmq port while
trying to start it to prevent interference
with client applications

3) Perform test of RMQ server start on promote

4) Do not check if we want to join the cluster -
simply join it - it is idempotent operation

5) Fix my_host() function determining if
our host is included into the list

6) Fix trim_var function to strip the line
instead of stripping the first argument

7) Stop slave node in case we failed
to join the master node. This will make
slave restart again and try to join again

8) Add debug option to monitor command

9) Add debug to several misc. functions

Change-Id: If5df451a6e2d72bf50c47c28d8a36b46045dd5cd