create keystone endpoints after keystone service
Keystone endpoints are created using python-keystoneclient which uses
Keystone API, which means that keystone service should be already
running. The dependency to run keystone db_sync before *::keystone::auth
is not strong enough, so it is replaced with dependency on keystone
service (which in turn is subscribed to keystone db_sync).

Since nova_floating_range depends on nova::keystone::auth, this will
also ensure that keystone service is running before floating IP ranges
are created with nova-network.

The wait-for-haproxy-nova-backend hack is removed: nova_floating_range
already depends on nova-api class (which declares nova-api service) and
on openstack::ha::nova class which reloads haproxy and declares nova-api
service dependency on keystone haproxy services (which in turn depend on
keystone service and also reload haproxy). With all these dependencies
in place, both keystone and nova services are running and registered
with haproxy before floating IP ranges are created.

Related-Bug: #1348171
Change-Id: Iad05b005fcc99d50507c36e934bce1fd1937ab6c