Restart containers on docker service refresh
Devicemapper bugs in docker 0.10 cause containers
to incorrectly restart when docker service is
restarted when Puppet upgrades the RPM. Added
lazy unmounts to dockerctl and a refreshonly
exec to start all containers when Docker service
is restarted (because of package upgrade).

Includes bugfixes for purging iptables rules and
starting nailgun and ostf containers while nginx
container is not yet started.

Change-Id: I9c6d9b04bdaa5cf1fcc9752cc1e8f42cf693de1b
Closes-Bug: #1358144