Clone neutron L3 and DHCP agents in corosync
Adjust the puppet manifests for the neutron l3 agents
and dhcp agents to set them to cloned services in corosync.
All of these agents should be running all of the time
since they are not sharing state or offering conflicting
services. Without this, a single agent will be a performance
bottleneck and a single point of failure for all tenant networks.

Closes-Bug: #1359082
Change-Id: I3b500a6268b391f7b63b50dc2a72952458226c69