Fix the Neutron DB creation during deployment
Neutron database should be created by the Neutron db migration scripts,
but they are ignored during deployment of HA clusters. The problem is
the migration scripts depends on neutron-server package installation,
but it's installed as a dependency for the OVS server package, so no
necessary event was produced by Puppet. Now it's installed separately.

Change-Id: I5d10ff9df6a37ad807d0d0214f3a6113bd58c08c
Closes-bug: #1361541