Run docker containers with host networking
Note: Requires Docker 1.2 or greater.

This change removes the need for dhcrelay
for cobbler container and the configuration
of masquerade ip forward rules. It will be
left installed in case of rollback to a
previous Fuel release.

Changes LXC driver to native exec driver
libcontainer. shell command now uses docker
exec command instead of lxc-attach.

Added privileged mode to nailgun container which
is necessary for sysctl to operate with host
networking mode.

Removed link from nailgun to mcollective because
it is not needed any longer.

blueprint fuel-master-docker-host-networking
blueprint fuel-master-docker-upgrade

Change-Id: I4111caf9ac7b33c514481704b3fb253966759748