Add x509_sign provider to openssl module.
It adds the x509_sign provider in order to be able to sign a CSR. It
adds two parameters that are: infile and template that is the config
file. Private key used by the CA to sign the certificate is passed
using the configuration file. You can use it as follow:

  x509_sign { "/tmp/signed_req.pem":
      ensure   => present,
      template => "/tmp/ca_openssl.cnf",
      infile   => "/tmp/user_req_cert.pem"
  }

Change-Id: I9be9df731be3bf64333eff366aaf9eb27a27b72d
Implements: blueprint ssl-endpoints