Added weighting parameters for docker containers
This feature adds LXC or libcontainer weighting constraints
to docker containers on Fuel Master. It is designed to be
compatible with both docker 0.10 and 1.x, at which time it
will be necessary to switch to libcontainer as the driver
engine. 
This sets the default CPU weight of each container to 1024
(LXC default) and gives 2x CPU priority to astute and cobbler,
and 4x priority to nailgun and mcollective containers. 
Memory limits are set to 5% of total system memory as a 
default, but 20% to mcollective and nailgun and 12.5% to
cobbler and astute. This will ensure that no single container
can occupy all CPU time or available memory.


blueprint improve-container-resource-management

Change-Id: If01fbe10624aa8685d9ae97508a6fd32082d2a7b