Add admin_net setting to nailgun-agent config
If a node has an admin IP address that is not the same subnet
as the Fuel master the nailgun-agent will fail to properly update
the node's IP after deployment. This leaves an invalid IP address in
the database which OSTF uses to connect to nodes and will cause
spurious test failures.

Change-Id: Ia623f9a038e994ccd65c036cc5a6eb1cdcd4551d