Do not kill swift replicator services
To move from exec to SysV provider of the service
in previous releases we had to kill swift replicator
service with swift-init kill command.

This is not what we need in 6.0 release as we
are already managing everything with default OS
providers. Moreover, this commit is error-prone
for Ubuntu with upstart provider.

Change-Id: I9ed3a030b4987f919a9f9ec2d9351f0b2c3623b0
Closes-bug: #1381018