Fix keystone_endpoint provider
Provider logic was broken for the new style keystone_endpoints
(the ones with "region/service" resource names). Due to this
puppet created new keystone endpoint even for existing ones on the
2nd+ runs.

We need to use common hash key for endpoint_hash in the provider
for both resource types ("region/service" and "service":region =>)
So we just force endpoint_hash to use "region/service" as hash key
for any "keystone_endpoint" resource type.

This patch and the problem will go away after we sync keystone
module with upstream.

Change-Id: I9eaaaf0f63db36a9837cc2141c293efd8adfd68e
Closes-bug: 1375254