Reduce dnsmasq verbosity and fix logrotate
dnsmasq dhcp logging is extremely verbose and can quickly
fill up logs for dnsmasq. It is not needed in most cases,
so it is now disabled.

Logrotate contained paths relative to rsyslog container,
but logrotate is actually run on Fuel Master host itself.
The paths were updated to be compatible for running
logrotate in a container or on the host.

Closes-Bug: #1382658
blueprint 100-nodes-support

Change-Id: I4290da4c8e3e5585396d105ce8e1f0cfceed61b3