Ensure qemu-kvm is running when deploying compute on KVM
qemu-kvm was not being started on ubuntu, which inserts
kvm kernel modules appropriately.
This behavior corrects a bug in qemu-kvm package on ubuntu
where it only loads kvm module after starting the qemu-kvm
service.
For CentOS, this functionality is provided in the script
/etc/sysconfig/modules/kvm.modules, and it is called just
to provide parity with Ubuntu, even though it does not
appear to be broken.

Change-Id: I05017f991593c79dbd6caa60683fd6061dd71626
Closes-Bug: #1383800