Optimize neutron-api waiting cycle
Currently there're 60 attempts to reach Neutron API with short delays between
them. It leads to very slow deployment failure in case of network problems,
because client's connection to server hangs for 60 seconds itself.
The idea is to decrease number of attempts but increase the delay, it should be
more optimal for different cases.

Change-Id: Icc4caa3c4eb3d22f24ff9ae7066ad1e60e65bb18
Related-bug: #1384332