Set iser_ip_address with a storage network address
The iser_ip_address parameter in /etc/cinder/cinder.conf
has a wrong value.

This parameter has to be configured with the storage
endpoint address, since it is no longer inherits the
iscsi_ip_address by default.

Currently it uses the $my_ip value, which is not capable
to work with RDMA transport (with its default value).

This fix assures that the default iser_ip_address will
use an address with RDMA support, if iSER is chosen as
the iSCSI transport by the user.

Closes-bug: #1387794

Change-Id: If0f67a9dbe80bfc8655dab85a96d3621bdaf38da