Improve wait-for-haproxy-* execs
Right after start haproxy marks all backends as UP. It marks
problem ones as DOWN only after health-checks. So we should not
rely on BACKEND status while INI is in process.

In order to optimize wait-for-haproxy-* execs we use new python
script that pulls Haproxy stat data and analyses it.

Change-Id: Ia3743d446826ee1a9d42799a83e14a911a577966
Related-bug: #1391180