Raise number of nailgun uWSGI workers and sockets
Fixes issue where nailgun stops responding to requests
because all workers are occupied and the queue of
available uWSGI sockets fills up by increasing the
number of workers (double) and raising the connection
limit from 100 to 4096.

This requires a kernel setting of net.core.somaxconn, which
is now being applied on Fuel Master host because it cannot
be done on an unprivileged Docker container.

Change-Id: I14724d5c6953570d53b6990ef0d6dccf07ddbaa4
Closes-Bug: #1391935