Specify the default floating IP pool in nova.conf
For now no default floating IP pool is specified in nova.conf 
when Neutron is deployed. This causes some problems. For example, 
when executing the command `nova floating-ip-create` we will have 
the following error:

ERROR (NotFound): Floating ip pool not found. (HTTP 404)

This issue happens only if floating IP pool is not provided to 
the command. The workaround is to provide pool to the command 
as an argument, but it is convenient to have the default pool 
and execute commands without extra arguments.

Closes-Bug: #1370448

Change-Id: I2c0f8659ed7c0afd925e2d468fa6019b1083e4a8