Create m1.micro nova flavor
m1.micro flavor contains 64MB RAM, 1vcpu, and no disk.

This enables users to deploy Cirros VM on compute nodes with
only 1GB memory for testing.

Note that this is implemented with an exec because of the lack of
nova flavor management in upstream puppet-nova, which will take
several months before it has novaclient support.

Change-Id: I827a14a96ecb3c1174b1e1a013f22bcb625afbfa
Closes-Bug: #1318512