Do additional checks in OCF only for action monitor.
W/o this change, the *monitor() function of OCF scripts
do additional AMQP connection checks unconditionally, every
time it has been invoked. And that could cause *too much* netstat
invocations for host OS to handle them w/o an issues.

By design, we want these additional checks performed only when
OCF script being invoked as a monitor action, but not as a start
or stop, for example.

The solution is to check $__OCF_ACTION inside of monitor() function
and act as appropriate. Also, a sleep 1 for action start loop should
be as well aligned with the other ones in other OCF scripts and it
should be 3 seconds instead.

Closes-bug: #1385755

Change-Id: I4eba9fd5e3eaa6448590808ea85bdf152b2ef176
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>