new option manage_password for keystone_user
Adds a new option manage_password, which defaults
to 'True' for all keystone_user objects, that
has the effect of enforcing the password of a given
keystone_user. If this is disabled, the user may change
his or her password at a later time and not be reset
by the keystone Puppet module.

Change-Id: I4bd59f233273374545953f16c7843488148096d6
Closes-Bug: #1400701