Add local clock source to ntp.conf
With new puppet NTP module we don't use local clock source for nodes,
if they are virtual. It's cause long ntp convergence time on master
node after fresh install and lead to impossibility to NTP sync on
bootstrap nodes. To fix that, we just remove check that node is virtual
(as it had done in old NTP puppet module).

Change-Id: I015cffe0e5df030e5d6142295ee27fc693b09cfa
Closes-Bug: 1298360