New provider for swift ringbuilder rebalance
WORK IN PROGRESS

The balance value is the highest percentage off the desired amount of
partitions a given device wants. For instance, if the "worst" device
wants (based on its weight relative to the sum of all the devices’
weights) 123 partitions and it has 124 partitions, the balance value
would be 0.83 (1 extra / 123 wanted * 100 for percentage).

So we should try to make sure balance is 0, because only 0 balance
means that all the devices got exact amount of partitions they
wanted to.

TODO:
 - write rspec tests

Change-Id: I917a3fbfb85b6c022b9d7234ee3c0437bb31f664
Closes-bug: #1405783