Switching to the default driver for Keystone policies
For now the driver keystone.policy.backends.rules.Policy is used
for Keystone policies. However such basic operations as create/list/
get/update/delete policy are not implemented in this driver. So we
need to use the default driver to be able to create/list/get/update/
delete Keystone policies.

Closes-Bug: #1418973

Change-Id: Ic0758fd5af87d968d2c7ff5a5fab8abc2a44246b