Rabbitmq reported not running in zabbix
RabbitMQ is reported as not running when rabbitmq is run
as root user. This patch removes the check as which
user rabbitmq is running.

This may introduce a security issue if our purpose was to
make sure that rabbitmq is always run as rabbitmq user.
If so please discard this patch and change init script
of non-ha mode to run RabbitMQ as rabbitmq user.

Change-Id: I7879d68d535880babc34d4321367592da94128f0
Closes-bug: #1419397