Add support for Neutron agents local state reports
It makes sense to detect additional Neutron agents problems like inability to
communicate with neutron server over message queue.
So in order to provide additional hints to monitoring scripts 'local reports'
are introduced. A local report is a piece of information written to a file
by neutron agent about status of certain kind of operations.
This patch adds support for them to Pacemaker OCF scripts.

This feature is experimental and is disabled by default.

Change-Id: Ie35ec805ce96e7a27fc699c9114c719d8cdcddef
Implements: blueprint neutron-agents-local-reports