Make one joined graph from 3 separate stages
Previously this 3 stages was disjoined and connected programatically
during traversal in nailgun. Such approach imposes some dificulties
during traversal and filtering and not really extandable (add new stages).

Introduced change allows us to simplify traversal logic, and therefore
improve UX for tasks developer

Also in such approach it is quite easy to introduce new stages, like
something that should be executed after cluster is completed, for example,
stage end_notification that should fire some messages when we can guarantee
that cluster is completed.
Or maybe we will want to introduce stage provision in the same graph.

related to blueprint granular-deployment-based-on-tasks

Change-Id: Ie332d0c1745a48c5ae96c4c80d94922c321d5f7a