Fix rabbitmq OCF get_status for action monitor
W/o this fix, RabbitMQ app running under load
sometimes can hang with action status. This is a
problem because Pacemaker cannot detect the resource
failure and keeps RabbitMQ app unresponsive.

The solution is to wrap with a timeout the rabbitmqctl
eval command invoked by get_status() so it would be
killed, if hanged.

Change-Id: I2b16cced07dd8ee4888b5ed2d66934bc601a845d
Closes-bug: #1423116
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>