Move admin user and openrc up so that its not relied on
The admin user defined in nailgun should not be used for any service
purpose. This pulls the user out of the lower levels of the manifests
so that it's harder to abuse this as a usage.

Commonly other modules have an their own admin {user,password,tenant}
parameters which should not be confused with this admin user which is
only for the enduser to consume.

Related-bug: #1396594
Related-bug: #1347542

Change-Id: Ie991a64617919727724638d582af301ee4519655