Remove openrc/admin-token usage in q-agent-cleanup.py
Removes the openrc/admin-token deps in q-agent-cleanup.py. Instead it will
read the neutron service credentials out of neutron.conf. This has the
advantage that it must be up to date for neutron to function so if the
credential is modified it will have to be updated by the administrator.

Related-bug: #1347542
Closes-bug: #1396594

Change-Id: Iee0968a079deacef24b56f34e9314b0ebecad0ae