Manage new compute services state
When deploying or scaling compute nodes,
disable all new compute services and use separate
post-deploy task to re-enable them back.
That should be done to prohibit schedulers to assign tasks
for these nodes until the deployment/scaling is done.

Nova-compute binary must be running in order to be first
seen by conductor, otherwise nova CLI would fail to disable
it. But nova-compute must be kept running for as
small amount of time as possible, because it shouldn't be
assigned for launching instances until the deployment is
done.
To achieve this, nova compute class is put in anchor and
deploys compute service not running. Next, a compute role
deployment task triggers compute service for start and
(almost) instantly disables it with nova CLI.
A post deploy task then starts nova-compute service
and enables it via nova CLI.

Security impact: displays nova service user password in
puppet logs.
Note, that this fix should be re-implemented later with
host-aggregates

DocImpact
Related blueprint disable-new-computes
Closes-bug: #1398817

Change-Id: Ia63d043753693360a008ec89924cdcdd93c007f3
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>