Manage new compute services state
When deploying or scaling compute nodes,
disable all new compute services and use separate
post-deploy task to re-enable them back.
That should be done to prohibit schedulers to assign tasks
for these nodes until the deployment/scaling is done.

Nova-compute binary must be running in order to be first
seen by conductor, otherwise nova CLI would fail to disable
it. But we also want to keep nova-compute running for as
small amount of time as possible, because we don't want it
to be assigned for launching instances.

So we deploy compute service not running, put nova compute
class in anchorm, then trigger compute service for start
and disable it with nova CLI as far as possible.

Note, that this fix should be reimplemented later with
host-aggregates.

DocImpact
Fuel-CI: disable
Related blueprint disable-new-computes
Closes-bug: #1398817

Change-Id: Ia63d043753693360a008ec89924cdcdd93c007f3
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>