Manage new compute services state
* When deploying or scaling compute nodes,
  disable all new compute services and use separate
  post-deploy task to re-enable them back.
* Restart nova-conductor service at primary controller after
  every compute task finished in order to apply to services
  status displayed by 'nova service-list' enable_new_services=False
  configuration change
  (Related bug https://bugs.launchpad.net/bugs/1426332)

That should be done to prohibit schedulers to assign tasks
for these nodes until the deployment/scaling is done.
For nova, this fix should be reimplemented later with host-aggregates.

Note, that if compute roles are shared with controllers,
we should keep new services enabled, otherwise all of the nova
services will be put in disabled state as well. So, the case with scaling
controllers with compute multiroles is not covered here.

DocImpact

Related blueprint disable-new-computes
Closes-bug: #1398817

Change-Id: Ia63d043753693360a008ec89924cdcdd93c007f3
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>