Manage new compute services state
When deploying or scaling an Openstack environment,
disable all of the nova-compute services and use a separate
post-deploy task to re-enable them back.
That should be done to prohibit nova schedulers to assign tasks
for compute nodes until the deployment or scaling is finished.

Note, that this fix may be re-implemented later with nova
host-aggregates

DocImpact
Related blueprint disable-new-computes
Closes-bug: #1398817

Change-Id: Ia63d043753693360a008ec89924cdcdd93c007f3
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>