Change default Ceph log file name to '$cluster-$name.log'
There is a possibility that an OSD host will run many OSD deamons, so it
is better idea to log messages to different files.

Note that '$cluster-$name.log' is a default value set by Ceph. In
result, radosgw messages are no longer saved in radosgw.log, but in
ceph-client.admin.log. The best solution is to modify daemon init script
to save this log under different name, but it would require us to
maintain them on our own, what seems no worth the hassle.

Change-Id: Ief8186006386af8ae7e40cffeeaeef5a5c0f3c70
Closes-Bug: 1428693