cluster: install pcs package on Debian/Ubuntu
The python-pcs package is the packet construction set [1]. This package
has been shipped with Debian/Ubuntu for quite a long time.
For some reason the pacemaker/corosync configuration system [2] has been
packaged for Ubuntu under the name `python-pcs'. Use the correct name
(that is, `pcs', like its RPM counterpart) to resolve the collision.

[1] https://github.com/gvnn3/PCS
[2] https://github.com/feist/pcs

Note: this patch should be merged along with the following ones:
https://review.fuel-infra.org/4522
https://review.openstack.org/164111

blueprint support-ubuntu-trusty

Change-Id: I460b550869d0fe0c4742c198ebc8c54b9548eae8