Remove memcached configuration from swift task
We configure memcached in a separate task so there's no need to do
the same in openstack::swift::proxy class, unless we moved swift to
the separate role.

Fix pre- post- tests for swift and memcached:
 * Add memcached process check for swift post test
 * Add testcase for memcached should not listen for public ip

Fix swift::proxy::cache parameter.

DocImpact: (Ops guide) if the memcached service was restarted, the
  keystone service must be restarted next as well.

Related-bug: #1432242
Change-Id: I5e965f12e5ba0003004ae10442d63c358a104367
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>