Setup swift rings rebalance and repush cronjobs
We should try to make sure the balance is 0, because only 0 balance
means that all the devices got exact amount of partitions they
wanted to and our ring is balanced.

In order to reduce network load we setup cronjobs that willrebalance
rings periodically and rsync them from primary controller to
secondaries.

Change-Id: I141c7cb581da2da0ef1b47cefaf5b9c485509ecd
Closes-bug: #1409783