Set amd64 architecture for dpkg
Since bp/separate-mos-from-linux the repo on the master node doesn't have
a package index for the i386 architecture. Meantime, by default Ubuntu
installer configures dpkg on slaves to use both amd64 and i386, which is
wrong because it's not intended.

This commit explicitly sets the architecture to use - 'amd64'.

DocImpact
Closes-Bug: #1436459
Change-Id: If8a0225c51cec8862b90f9cc62e660b08b41edf8