Increase deadlocks tolerance for galera
The default value for wsrep_retry_autocommit, which is 1,
does not provide good tolerance for the deadlocks.

This is a problem, if an application does not tolerate
deadlocks well and there are write-intensive loads and
1 retry could be not enough to wait for the slow
transaction to end and the lock to gone.

The solution is to increase an amount of retries what may
help to hide deadlock errors.

Related-bug: #1431702

Change-Id: Ic7dd5868522f790ef31c193165a4ec6ee80436e3
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>