Set MySQL transaction isolation to READ_COMMITED
A MySQL uses repeatable read for transactions by default.
A Postgres uses read committed by default, and an SQLite
uses serializable. Oslo.db don't set the isolation level
explicitly anywhere.

This is a problem as Openstack applications using Oslo.db
may be running under different isolation levels depending
on the backend type but the implicit isolation level
is unknown for Oslo.db code.

The solution is that Fuel should configure the MySQL
transaction_isolation=READ-COMMITTED as it looks that this
mode is implicit for backends by Oslo.db because it is a
standard for DB transations.

Change-Id: Ib4d22e1cdfb177ce236a89e7a6292292018052fa
Closes-bug: #1438107
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>