Do not re-elect RabbitMQ multistate resource master
Current monitor action for multistate RabbitMQ resource
in Pacemaker  may re-elect existing master when there is
no need to do so. That is a problem as master elections
introduce full cluster downtime and we don't want
additional elections.

The fix is to check whether the given node uptime is equal
to the other nodes uptime and drop this node from the list
of candidates if there is a master exists. Note, there is a
special 'rabbit-master' attribute in CIB for this.

Closes-bug: #1435254

Change-Id: I6f2484fb8f3284f76461c184148d27aa86a5d4b6
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>