Fix post-start notify exit code for rabbit OCF
There is an error in OCF logic causing the resource to be reported
as $OCF_NOT_RUNNING that leads to additional restarts for rabbitmq
resources in pacemaker.

The solution is to ensure that only the node which actually has
started and is joining the cluster may fail with this error code,
while the other nodes may not.

Closes-bug: #1438699

Change-Id: I8d3b6e8f76a6a89608e59a52081c21931e5654fb
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>