Set mtu for internal networks according to segmentation type
Currently MTU for Neutron private networks is always set to 1500,
which is not correct for GRE segmentation and causes performance degradation.
MTU value should also depend on MTU settings of physical interfaces.
The only way to change MTU for Neutron private network is to use an additional
dnsmasq configuration file.

Co-Authored-By: Sergey Vasilenko <svasilenko@mirantis.com>
Change-Id: Iaa9631b58f268ec414c58d1b7c47de3573772fc5
Closes-bug: #1411732