Pmanager: Add udev settle after trigger
Since udev is well known for its async nature, we need to be sure
that it finishes its queue processing after calling 'udevadm trigger'.
'udevadm settle' is the only way to achive that sync.

Otherwise, we could end up with unexpectable changes in devfs in the
middle of the rest of the script execution.

Change-Id: Ie994b5fbb070314621ba369e7a4f4850cf400173
Related-Bug: #1410471