Make block device mapping timeout configurable
When booting instances passing in block-device and increasing the
volume size, instances can go in to error state if the volume takes
longer to create than 3 minutes (default value).

This change adds support for new cluster configutation option:
- block_device_allocate_timeout

And defaults it to 900 (15 minutes). This option is configurable
via fuel cli or Hiera.

DocImpact

Change-Id: If9148da046d31aab60a13bd70e94700331770000
Closes-bug: #1280399