Disable resources cleanup for DHCP agent by default
Now all resources created by DHCP agent (namespaces, processes) are removed
on agent start and stop. This operation can be rather slow when many networks
exist in a cloud, and timeout will turn the resource to unmanaged state.
To avoid such cases resources are not deleted by default.
It also helps to avoid service interruption because all resource have to be
re-created by the agent in case of restart.

Change-Id: Ia6cdc3b86b162f4614a8ed93838d3e09d9126ffc
Closes-bug: #1436414