Fix RabbitMQ apps eval in OCF
W/o this fix, if there are no apps running and rabbit node is actually
not functioning, get_status() would still report 0 considering the rabbit
resource is running. This is an issue as it may lead to the situations
when the resource reported OK, but in fact, the rabbit node is not a cluster
member.

The solution is to not rely only on which_applications() eval exit code
and test if the kernel app is running. Otherwise consider the pacemaker
resource is "not running" as well.

Closes-bug: #1446251

Change-Id: Ia2fcb18abb3d977c5fcb26bfdeac864b6834f478
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>
(cherry picked from commit bc4a7ec8093db81d8d1de478788825e228890b05)